Read for fixes
The creeper feels more powerful than other weapons around its tier. Maybe decrease its damage by 2 (DONE)
Putrid spitter shouldn't drop itself. It doesn't make sense, since it "explodes" (DONE)
Freeman needs a slight nerf. Lower its damage or increase its usetime slightly (DONE)
Freeman shouldn't drop every time from the mecromancer. Either create a loot pool or make it a rarer drop (DONE)
Make the dust effect of the coiled energizer a little less intense (around the coiled energizer itself I mean), Maybe reduce the dust and the distance away the dust can spawn slightly
Make the head of the coiled rod projectile "light up" a bit before and as it attacks (DONE)
Coiled energizer should require a crafting station
It might just be the prefix I have, but floran hamaxe mining and swinging speeds aren't synced (DONE)
Vinewrath himself despawned when i got too far away (i didn't even leave the reach)
Vinewrath bane fight drags on a bit too long. Maybe reduce his first phase health by 10%, and up his damage to make up for that? (DONE)
That one bright green dust projectile the vinewrath shoots in his final phase should despawn when he dies. It chased me for a pretty second after his death (DONE)
New sepulchre tiles
Mossy Stone for the reach that doesn't get overriden
Moved Blood Moon enemies to a new folder
Buffed Coconut
Deleted Bloodfire Staff
Made recommended nerfs and buffs
Added space junk (incomplete)
Removed marble ore spawning in Asteroids
Added many gorees

+++ Projectiles/Hostile/CoilRocket.cs
@@ -31,6 +31,7 @@ namespace SpiritMod.Projectiles.Hostile
 			projectile.ignoreWater = true;
 			projectile.aiStyle = -1;
 		}
+        int numHits;
         public override bool PreAI()
         {
             var list = Main.projectile.Where(x => x.Hitbox.Intersects(projectile.Hitbox));
@@ -38,7 +39,11 @@ namespace SpiritMod.Projectiles.Hostile
             {
                 if (projectile != proj && proj.friendly)
                 {
-                    projectile.Kill();
+                    numHits++;
+                    if (numHits >= 3)
+                    {
+                        projectile.Kill();
+                    }
                 }
             }
             return true;
